Finish translator refactor
