+++ packages/xlucene-evaluator/src/index.ts
@@ -1,15 +1,13 @@
 import DocumentMatcher from './document-matcher';
 import LuceneQueryParser from './lucene-query-parser';
 import LuceneQueryAccess, { QueryAccessConfig } from './lucene-query-access';
-import Translator, { ElasticsearchDSLResult } from './translator';
 
 export * from './interfaces';
 export * from './utils';
+export * from './translator';
 export {
     DocumentMatcher,
     LuceneQueryParser,
     LuceneQueryAccess,
     QueryAccessConfig,
-    Translator,
-    ElasticsearchDSLResult
 };
